NeoVim: Massive reduction in configuration
Removed any configuration that I didn't immediately recognize the purpose of,
and any options and/or plugins that I have not used in a long time. This
removed most all of the configuration.

I am trying to get back to as basic as I can in all the major programs I use.
This is merely the beginning. BEWARE THE CULL :3